requirements.txt
cffi
Cython
GitPython
json2html
lxml
pyOpenSSL
requests
pyjnius
pytest
pytest-cover